requirements.txt
flask
faiss-cpu
sentence-transformers
google-genai
numpy
python-dotenv